+++ boolean-logic-ultimate/FilterUltimate.js
@@ -13,20 +13,24 @@ module.exports = function (RED) {
 		setNodeStatus({ fill: "grey", shape: "dot", text: "Waiting" });
 		this.on('input', function (msg) {
 
+			const utils = require("./utils.js");
+			let sPayload = utils.fetchFromObject(msg, config.payloadPropName || "payload");
+
 			// 15/11/2021 inform user about undefined topic or payload
-			if (!msg.hasOwnProperty("payload") || msg.payload === undefined || msg.payload === null) {
+			if (sPayload=== undefined ) {
 				setNodeStatus({ fill: "red", shape: "dot", text: "Received invalid payload from " + msg.topic || "" });
 				return;
 			}
 
+			
 			var bRes = null;
 			try {
-				bRes = ToBoolean(msg.payload);
+				bRes = utils.ToBoolean(sPayload);
 			} catch (error) {
 			}
 
 			if (bRes === undefined || bRes === null) {
-				setNodeStatus({ fill: "red", shape: "dot", text: "Received non convertible boolean value " + msg.payload + " from " + msg.topic });
+				setNodeStatus({ fill: "red", shape: "dot", text: "Received non convertible boolean value " + sPayload + " from " + msg.topic });
 				return;
 			}
 
@@ -44,31 +48,6 @@ module.exports = function (RED) {
 
 		});
 
-
-
-		function ToBoolean(value) {
-			let res = false;
-			let decimal = /^\s*[+-]{0,1}\s*([\d]+(\.[\d]*)*)\s*$/
-
-			if (typeof value === 'boolean') {
-				res = value;
-			}
-			else if (typeof value === 'number' || typeof value === 'string') {
-
-				if (typeof value === "string" && value.toLowerCase() === "on") return true;
-				if (typeof value === "string" && value.toLowerCase() === "off") return false;
-
-				// Is it formated as a decimal number?
-				if (decimal.test(value)) {
-					res = parseFloat(value) != 0;
-				}
-				else {
-					res = value.toLowerCase() === "true";
-				}
-			}
-
-			return res;
-		};
 	}